Legacy support for Voigt function
