+++ hyperspy/_components/voigt.py
@@ -16,133 +16,92 @@
 # You should have received a copy of the GNU General Public License
 # along with  HyperSpy.  If not, see <http://www.gnu.org/licenses/>.
 
-import numpy as np
 import math
+import numpy as np
+import sympy
 
-from hyperspy.component import Component
+from hyperspy._components.expression import Expression
+from distutils.version import LooseVersion
 
 sqrt2pi = math.sqrt(2 * math.pi)
+sigma2fwhm = 2 * math.sqrt(2 * math.log(2))
 
+class Voigt(Expression):
 
-def voigt(x, FWHM=1, gamma=1, center=0, scale=1):
-    """Voigt lineshape.
-
-    The voigt peak is the convolution of a Lorentz peak with a Gaussian peak.
-
-    The formula used to calculate this is::
-
-        z(x) = (x + 1j gamma) / (sqrt(2) sigma)
-        w(z) = exp(-z**2) erfc(-1j z) / (sqrt(2 pi) sigma)
-
-        V(x) = scale Re(w(z(x-center)))
-
+    r"""Voigt component.
+    
+    Symmetric peak shape based on the convolution of a Lorentzian and Normal 
+    (Gaussian) distribution. In this case using an approximate formula by David 
+    (see Notes).
+    
+    .. math:: 
+        z(x) &= \frac{x + i \gamma}{\sqrt{2} \sigma} \\
+        w(z) &= \frac{e^{-z^2} \text{erf}(-i z)}{\sqrt{2 \pi} \sigma} \\
+        f(x) &= A \Re\left\{ w \left[ z(x-x_0) \right] \right\}
+    
+    
+    ============== =============
+    Variable        Parameter 
+    ============== =============
+    :math:`x_0`     centre 
+    :math:`A`       area
+    :math:`\gamma`  gamma
+    :math:`\sigma`  sigma
+    ============== =============
+    
+    
     Parameters
-    ----------
-    gamma : real
-       The half-width half-maximum of the Lorentzian
-    FWHM : real
-       The FWHM of the Gaussian
-    center : real
-       Location of the center of the peak
-    scale : real
-       Value at the highest point of the peak
-
+    -----------
+    centre : float
+        Location of the maximum of the peak.
+    area : float
+        Intensity below the peak.
+    lwidth : float
+        HWHM=:math:`\gamma` of the Lorentzian distribution.
+    gwidth: float
+        FWHM=:math:`2 \sigma \sqrt(2 \log(2))` of the Gaussian distribution.
+    
+    For convenience the `sigma` and `gamma` attributes can also be used to set 
+    set and get the width of the Gaussian and Lorentzian parts of the 
+    distribution, respectively. The `height` attribute can be used to get and 
+    set the height of the distribution.
+        
     Notes
     -----
-    Ref: W.I.F. David, J. Appl. Cryst. (1986). 19, 63-64
-
-    adjusted to use stddev and HWHM rather than FWHM parameters
-
-    """
-    # wofz function = w(z) = Fad[d][e][y]eva function = exp(-z**2)erfc(-iz)
-    from scipy.special import wofz
-    sigma = FWHM / 2.3548200450309493
-    z = (np.asarray(x) - center + 1j * gamma) / (sigma * math.sqrt(2))
-    V = wofz(z) / (math.sqrt(2 * np.pi) * sigma)
-    return scale * V.real
-
-
-class Voigt(Component):
-
-    """Voigt profile component with support for shirley background,
-    non_isochromaticity, transmission_function corrections and spin orbit
-    splitting specially suited for Photoemission spectroscopy data analysis.
-
-    f(x) = G(x)*L(x) where G(x) is the Gaussian function and L(x) is the
-    Lorentzian function
-
-    Attributes
-    ----------
-
-    area : Parameter
-    centre: Parameter
-    FWHM : Parameter
-    gamma : Parameter
-    resolution : Parameter
-    shirley_background : Parameter
-    non_isochromaticity : Parameter
-    transmission_function : Parameter
-    spin_orbit_splitting : Bool
-    spin_orbit_branching_ratio : float
-    spin_orbit_splitting_energy : float
-
+    W.I.F. David, J. Appl. Cryst. (1986). 19, 63-64
     """
 
-    def __init__(self):
-        Component.__init__(self, (
-            'area',
-            'centre',
-            'FWHM',
-            'gamma',
-            'resolution',
-            'shirley_background',
-            'non_isochromaticity',
-            'transmission_function'))
-        self._position = self.centre
-        self.FWHM.value = 1
-        self.gamma.value = 0
-        self.area.value = 1
-        self.resolution.value = 0
-        self.resolution.free = False
-        self.shirley_background.free = False
-        self.non_isochromaticity.value = 0
-        self.non_isochromaticity.free = False
-        self.transmission_function.value = 1
-        self.transmission_function.free = False
-        # Options
-        self.shirley_background.active = False
-        self.spin_orbit_splitting = False
-        self.spin_orbit_branching_ratio = 0.5
-        self.spin_orbit_splitting_energy = 0.61
+    def __init__(self, centre=0., area=1., lwidth=0.2, gwidth=0.05, 
+                 module="scipy", **kwargs):
+        # Not to break scripts once we remove the legacy Polynomial
+        if "legacy" in kwargs:
+            del kwargs["legacy"]
+        if LooseVersion(sympy.__version__) < LooseVersion("1.3"):
+            raise ImportError("The `Voigt` component requires "
+                              "SymPy >= 1.3")
+        super(Voigt, self).__init__(
+            expression="area * real(V); \
+                V = wofz(z) / (sqrt(2.0 * pi) * sigma); \
+                z = (x - centre + 1j * lwidth) / (sigma * sqrt(2.0)); \
+                sigma = gwidth / (2 * sqrt(2 * log(2)))",
+            name="Voigt",
+            centre=centre,
+            area=area,
+            lwidth=lwidth,
+            gwidth=gwidth,
+            module=module,
+            autodoc=False,
+            **kwargs,
+        )
+
+        # Boundaries
+        self.area.bmin = 0.
+        self.lwidth.bmin = 0.
+        self.gwidth.bmin = 0.
 
         self.isbackground = False
         self.convolved = True
 
-    def function(self, x):
-        area = self.area.value * self.transmission_function.value
-        centre = self.centre.value
-        ab = self.non_isochromaticity.value
-        if self.resolution.value == 0:
-            FWHM = self.FWHM.value
-        else:
-            FWHM = math.sqrt(self.FWHM.value ** 2 + self.resolution.value ** 2)
-        gamma = self.gamma.value
-        k = self.shirley_background.value
-        f = voigt(x,
-                  FWHM=FWHM, gamma=gamma, center=centre - ab, scale=area)
-        if self.spin_orbit_splitting is True:
-            ratio = self.spin_orbit_branching_ratio
-            shift = self.spin_orbit_splitting_energy
-            f2 = voigt(x, FWHM=FWHM, gamma=gamma,
-                       center=centre - ab - shift, scale=area * ratio)
-            f += f2
-        if self.shirley_background.active:
-            cf = np.cumsum(f)
-            cf = cf[-1] - cf
-            self.cf = cf
-            return cf * k + f
-        else:
-            return f
 
     def estimate_parameters(self, signal, E1, E2, only_current=False):
         """Estimate the Voigt function by calculating the momenta of the 
@@ -214,7 +173,7 @@ class Voigt(Component):
         height = data.max(i)
         if only_current is True:
             self.centre.value = center
-            self.FWHM.value = sigma * 2.3548200450309493
+            self.gwidth.value = sigma
             self.area.value = height * sigma * sqrt2pi
             return True
         else:
@@ -222,8 +181,32 @@ class Voigt(Component):
                 self.create_arrays(signal.axes_manager.navigation_shape)
             self.area.map['values'][:] = height * sigma * sqrt2pi
             self.area.map['is_set'][:] = True
-            self.FWHM.map['values'][:] = sigma * 2.3548200450309493
-            self.FWHM.map['is_set'][:] = True
+            self.gwidth.map['values'][:] = sigma
+            self.gwidth.map['is_set'][:] = True
             self.centre.map['values'][:] = center
             self.centre.map['is_set'][:] = True
             return True
+
+    @property
+    def sigma(self):
+        return self.gwidth.value / sigma2fwhm
+
+    @sigma.setter
+    def sigma(self, value):
+        self.gwidth.value = value * sigma2fwhm
+
+    @property
+    def gamma(self):
+        return self.lwidth.value
+
+    @gamma.setter
+    def gamma(self, value):
+        self.lwidth.value = value
+        
+    @property
+    def height(self):
+        return self.area.value / (self.gwidth.value * sqrt2pi)
+
+    @height.setter
+    def height(self, value):
+        self.area.value = value * self.gwidth.value * sqrt2pi
